Refactor shortcut decoding into an extra file.
Adding a dict `shortcut_object_decoders` that contains functions
that handles decoding of specific objects quicker than the generic method.

remove requirements.txt
