--- requirements.txt
@@ -1,12 +1,0 @@
-colcon-ros==0.3.14
-colcon-common-extensions==0.2.1
-lark-parser==0.8.1
-numpy==1.18.1
-flake8==3.7.9
-pydocstyle==5.0.2
-Cython==0.29.14
-lxml==4.5.0
-cryptography==2.8
-tornado==6.0.3
-pymongo==3.10.1
-Pillow==7.0.0
